+++ Angular-Frontend/src/app/User/deleted-users/deleted-users.component.ts
@@ -5,8 +5,8 @@ import {UserService} from '../../model/user.service';
 import {TitleService} from '../../title.service';
 import {Router} from '@angular/router';
 import {NgxSpinnerService} from 'ngx-spinner';
-import {AuthService} from "../../model/auth.service";
-import {SessionClass} from "../../model/session";
+import {AuthService} from '../../model/auth.service';
+import {SessionClass} from '../../model/session';
 
 @Component({
   selector: 'app-deleted-users',
@@ -29,8 +29,8 @@ export class DeletedUsersComponent implements OnInit {
   ngOnInit(): void {
     this.user = new User();
     this.appService.setTitle('Deleted Users');
-    this.deleteOld(1);
     this.reloadData();
+    // this.deleteOld(1);
   }
 
   restoreUser(id: number) {
@@ -46,7 +46,7 @@ export class DeletedUsersComponent implements OnInit {
               setTimeout(() => {
                 this.SpinnerService.hide();
                 this.reloadData();
-              }, 500);
+              }, 600);
             });
         });
   }
@@ -61,7 +61,7 @@ export class DeletedUsersComponent implements OnInit {
     this.userService.getUserList()
       .subscribe(data => {
         while (data[counter] != null) {
-          if (data[counter].userDeleted != null) {
+          if (data[counter].userDeleted != null && data[counter].network.netName === this.info.network) {
             this.userService.deleteUser(data[counter].userId).subscribe();
           }
           counter++;
